refactor(vmm): move the shared VM config types into mvm-vmm::host
base/config.rs is 297 lines of constants and plain types with zero
crate:: dependencies — a pure leaf. It was the last thing keeping the
Firecracker observe/guards/control modules and firecracker.rs pinned to
mvm-runtime, since all four read VmSlot and the FC network/path
constants from it.

Moved rather than duplicated, and re-exported from mvm-runtime::base so
both mvm-runtime and mvm-hostd keep resolving it at the old path.
